WIP: Group creation and single ticket deletion

+++ src/main/java/assertionbit/trainapi/repositories/TicketRepository.java
@@ -5,9 +5,9 @@ import assertionbit.trainapi.jooq.public_.tables.RouteTickets;
 import assertionbit.trainapi.jooq.public_.tables.Tickets;
 import assertionbit.trainapi.jooq.public_.tables.TicketsGroup;
 import org.jooq.DSLContext;
-import org.jooq.Record;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -54,7 +54,12 @@ public class TicketRepository {
                 .toList();
     }
 
-    public void addTicket(TicketEntity ticketEntity, Long routeId, Long sitId) {
+    public Long addTicket(
+            TicketEntity ticketEntity,
+            Long routeId,
+            Long sitId,
+            Long groupId
+    ) {
         var id = context.insertInto(tickets,
                         tickets.CREATION_DATE)
                 .values(Collections.singleton(ticketEntity.getCreation_date()))
@@ -65,12 +70,19 @@ public class TicketRepository {
         ticketEntity.setId(id);
 
         context.insertInto(routeTickets,
-                routeTickets.ROUTE_ID, routeTickets.SIT_ID, routeTickets.TICKET_ID)
+                routeTickets.ROUTE_ID,
+                routeTickets.SIT_ID,
+                routeTickets.TICKET_ID,
+                routeTickets.GROUP_ID)
                 .values(
                         Math.toIntExact(routeId),
                         Math.toIntExact(sitId),
-                        Math.toIntExact(ticketEntity.getId()))
+                        Math.toIntExact(ticketEntity.getId()),
+                        groupId == null ? null : Math.toIntExact(groupId)
+                )
                 .execute();
+
+        return id;
     }
 
     public List<Integer> getTicketRoute(Long ticketId) {
@@ -80,16 +92,34 @@ public class TicketRepository {
                 .map(s -> (Integer) s.get("route_id"));
     }
 
-    public void deleteTicketGroup(Long id) {
-    }
-
-    public void deleteTicket(Long id) {
-        context.delete(routeTickets)
-                .where(routeTickets.TICKET_ID.eq(Math.toIntExact(id)))
-                .execute();
+    public boolean deleteTicket(Long id) {
+        var affected = context.delete(routeTickets)
+                .where(
+                    routeTickets.TICKET_ID.eq(Math.toIntExact(id)),
+                    routeTickets.GROUP_ID.isNull()
+                ).execute();
+        if(affected == 0) {
+            return false;
+        }
         context.delete(tickets)
                 .where(tickets.ID.eq(Math.toIntExact(id)))
                 .execute();
+        return true;
+    }
+
+    public void addGroupTickets(
+            Long routeId,
+            ArrayList<Integer> sits
+    ) {
+        Long id = createGroupTicket();
+        for(int i = 0; i < sits.size(); i++) {
+            addTicket(
+                    new TicketEntity(LocalDateTime.now()),
+                    routeId,
+                    Long.valueOf(sits.get(0)),
+                    id
+            );
+        }
     }
 
     public List<Integer> getGroupOfTicket(Long id) {
@@ -99,5 +129,17 @@ public class TicketRepository {
                 .where(routeTickets.TICKET_ID.eq(Math.toIntExact(id)))
                 .fetch()
                 .map(s -> (Integer) s.get("group_id"));
+    }
+
+    public void deleteTicketGroup(Long id) {}
+
+    protected Long createGroupTicket() {
+        return this.context
+                .insertInto(ticketsGroup, ticketsGroup.CREATION_DATE)
+                .values(LocalDateTime.now())
+                .returningResult(ticketsGroup.ID)
+                .fetchOne()
+                .into(long.class);
+
     }
 }
